requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
Pillow==10.0.1
bcrypt==4.0.1
Flask-Session==0.5.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-socketio==5.8.0
eventlet==0.33.3
pytz==2023.3
python-telegram-bot==20.7
redis==5.0.1 